+++ linkis-commons/linkis-rpc/src/main/java/org/apache/linkis/rpc/message/annotation/Method.java
@@ -15,13 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.linkis.rpc.serializer.annotation;
+package org.apache.linkis.rpc.message.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target({ElementType.METHOD})
+@Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Receiver {}
+public @interface Method {
+    String value() default "";
+}
