+++ linkis-computation-governance/linkis-engineconn-manager/linkis-engineconn-manager-server/src/main/scala/org/apache/linkis/ecm/server/service/impl/DefaultEngineConnPidCallbackService.scala
@@ -22,7 +22,7 @@ import org.apache.linkis.ecm.server.LinkisECMApplication
 import org.apache.linkis.ecm.server.listener.EngineConnPidCallbackEvent
 import org.apache.linkis.ecm.server.service.EngineConnPidCallbackService
 import org.apache.linkis.governance.common.protocol.task.ResponseEngineConnPid
-import org.apache.linkis.rpc.serializer.annotation.Receiver
+import org.apache.linkis.rpc.message.annotation.Receiver
 
 
 class DefaultEngineConnPidCallbackService extends EngineConnPidCallbackService with Logging {
