+++ linkis-computation-governance/linkis-manager/linkis-application-manager/src/main/scala/org/apache/linkis/manager/am/service/engine/DefaultEngineCreateService.scala
@@ -45,7 +45,7 @@ import org.apache.linkis.manager.label.service.{NodeLabelService, UserLabelServi
 import org.apache.linkis.manager.label.utils.LabelUtils
 import org.apache.linkis.manager.persistence.{NodeManagerPersistence, NodeMetricManagerPersistence, ResourceManagerPersistence}
 import org.apache.linkis.manager.service.common.label.{LabelChecker, LabelFilter}
-import org.apache.linkis.rpc.serializer.annotation.Receiver
+import org.apache.linkis.rpc.message.annotation.Receiver
 import org.apache.linkis.resourcemanager.service.ResourceManager
 import org.apache.linkis.resourcemanager.{AvailableResource, NotEnoughResource}
 import org.apache.linkis.rpc.Sender
